requirements.txt
boto3==1.6.21
botocore==1.9.21
click==6.7
docutils==0.14
Flask==0.12.2
futures==3.2.0
gunicorn==19.7.1
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
python-dateutil==2.6.1
s3transfer==0.1.13
six==1.11.0
Werkzeug==0.14.1
python-memcached==1.59